Feature/profile

+++ app/Http/Controllers/Account/AccountCompanyController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Account;
 
+use App\Dto\CompanyDto as CompanyFrontend;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Company\CreateCompanyRequest;
 use App\Http\Requests\Company\DeleteCompanyRequest;
@@ -11,6 +12,14 @@ use App\Http\Resources\CompanyResource;
 use App\Http\Responses\JsonApiResponse;
 use App\Http\Responses\JsonErrorResponse;
 use App\Services\Company\CompanyService;
+use App\Services\Company\Dto\CompanyDto;
+use App\Services\Department\DepartmentService;
+use App\Services\Department\Dto\CreateDepartmentDto;
+use App\Services\Permission\PermissionService;
+use App\Services\Profile\Dto\CreateProfileDto;
+use App\Services\Profile\ProfileService;
+use App\Services\Role\Dto\CreateRoleDto;
+use App\Services\Role\RoleService;
 use App\Services\User\UserService;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -22,7 +31,11 @@ class AccountCompanyController extends Controller
 {
     public function __construct(
         private readonly UserService $userService,
-        private readonly CompanyService $companyService
+        private readonly CompanyService $companyService,
+        private readonly RoleService $roleService,
+        private readonly DepartmentService $departmentService,
+        private readonly PermissionService $permissionService,
+        private readonly ProfileService $profileService,
     ) {
     }
 
@@ -32,7 +45,12 @@ class AccountCompanyController extends Controller
 
         $companiesIds = $this->userService->getCompanyIds($userId);
 
-        $companies = $this->companyService->findMany($companiesIds);
+        $dbCompanies = $this->companyService->findMany($companiesIds);
+
+        $companies = [];
+        foreach ($dbCompanies as $company) {
+            $companies[] = $this->formatCompanyDtoToFrontend($company);
+        }
 
         $companyCollection = CompanyResource::collection($companies);
 
@@ -42,14 +60,52 @@ class AccountCompanyController extends Controller
     public function store(CreateCompanyRequest $request)
     {
         $userId = Uuid::fromString(Auth::id());
+        $user = $this->userService->find($userId);
 
         $createCompanyDto = $request->getDto();
-
         $company = $this->companyService->create($createCompanyDto);
 
-        //TODO profileService create profile in company
+        $this->userService->appendCompanyToUser($company->getUuid(), $user->getId());
 
-        return new JsonApiResponse((CompanyResource::make($company))->toArray($request));
+
+        $createDepartmentDto = new CreateDepartmentDto(
+            $company->getUuid(),
+            "Администрация"
+        );
+        $department = $this->departmentService->createDepartmentByCompanyId($createDepartmentDto);
+
+
+        $createRoleDto = new CreateRoleDto(
+            $company->getUuid(),
+            "Генеральный Директор"
+        );
+        $role = $this->roleService->createRoleByCompanyId($createRoleDto);
+
+
+        $dbPermissions = $this->permissionService->all();
+        $permissionsIds = [];
+        foreach ($dbPermissions as $permission) {
+            $permissionsIds[] = $permission->getUuid()->toString();
+        }
+        $this->roleService->appendPermissionsToRole($role->getId(), $permissionsIds);
+
+
+        $createProfileDto = new CreateProfileDto(
+            userId: $user->getId(),
+            companyId: $company->getUuid(),
+            roleId: $role->getId(),
+            departmentId: $department->getId(),
+            firstName: $user->getFirstName(),
+            lastName: $user->getLastName(),
+            secondName: $user->getSecondName(),
+            phone: $user->getPhone(),
+            photo: $user->getPhoto(),
+            email: $user->getEmail(),
+            birthday: $user->getBirthday(),
+        );
+        $profile = $this->profileService->createProfile($createProfileDto);
+
+        return new JsonApiResponse((CompanyResource::make($this->formatCompanyDtoToFrontend($company)))->toArray($request));
     }
 
     public function show(ShowCompanyRequest $request)
@@ -63,7 +119,7 @@ class AccountCompanyController extends Controller
 
         $company = $this->companyService->find($companyId);
 
-        return new JsonApiResponse((CompanyResource::make($company))->toArray($request));
+        return new JsonApiResponse((CompanyResource::make($this->formatCompanyDtoToFrontend($company)))->toArray($request));
     }
 
     public function update(UpdateCompanyRequest $request)
@@ -74,11 +130,16 @@ class AccountCompanyController extends Controller
         if (!$exists) {
             return new JsonErrorResponse(__('errors.company.not exists'), status: Response::HTTP_FORBIDDEN);
         }
-        //TODO profileService update check
+
+        $user = $this->userService->firstUserInCompany($updateCompanyDto->getUuid());
+
+        if($user->getId() != Auth::id()) {
+            return new JsonErrorResponse(__('errors.company.update'), status: Response::HTTP_FORBIDDEN);
+        }
 
         $company = $this->companyService->update($updateCompanyDto);
 
-        return new JsonApiResponse((CompanyResource::make($company))->toArray($request));
+        return new JsonApiResponse((CompanyResource::make($this->formatCompanyDtoToFrontend($company)))->toArray($request));
     }
 
     public function destroy(DeleteCompanyRequest $request)
@@ -89,7 +150,12 @@ class AccountCompanyController extends Controller
         if (!$exists) {
             return new JsonErrorResponse(__('errors.company.not exists'), status: Response::HTTP_FORBIDDEN);
         }
-        //TODO profileService destroy check
+
+        $user = $this->userService->firstUserInCompany($companyId);
+
+        if($user->getId() != Auth::id()) {
+            return new JsonErrorResponse(__('errors.company.update'), status: Response::HTTP_FORBIDDEN);
+        }
 
         $result = $this->companyService->delete($companyId);
 
@@ -115,5 +181,20 @@ class AccountCompanyController extends Controller
         }
 
         return $exists;
+    }
+
+    private function formatCompanyDtoToFrontend(CompanyDto $companyDto): CompanyFrontend
+    {
+        return new CompanyFrontend(
+            uuid: $companyDto->getUuid(),
+            name: $companyDto->getName(),
+            inn: $companyDto->getInn(),
+            kpp: $companyDto->getKpp(),
+            email: $companyDto->getEmail(),
+            phone: $companyDto->getPhone(),
+            url: $companyDto->getUrl(),
+            description: $companyDto->getDescription(),
+            isActive: $companyDto->isActive(),
+        );
     }
 }
